move files to new folders
